+++ src/libtrx/include/libtrx/config/types.h
@@ -134,7 +134,7 @@ typedef struct {
             BAR_SHOW_MODE show_mode;
             BAR_LOCATION location;
             BAR_COLOR color;
-        } lara_health_bar, lara_air_bar, lara_sprint_bar;
+        } lara_health_bar, lara_air_bar, lara_sprint_bar, lara_exposure_bar;
         struct {
             BAR_SHOW_MODE show_mode;
             BAR_LOCATION location;